Fix delete button on Main Window

+++ MusicPlayApp/MusicPlayList/MainWindow.xaml.cs
@@ -466,11 +466,11 @@ namespace MusicPlayList
 
                         if (result == MessageBoxResult.Yes)
                         {
-                            _songService.RemoveSongAsync(selectedSong);  // Xóa bài hát khỏi cơ sở dữ liệu
+                            await _songService.RemoveSongAsync(selectedSong);  // Xóa bài hát khỏi cơ sở dữ liệu
 
                             // Tải lại Playlist và Favorite List để cập nhật thay đổi
                             playlistListBox.ItemsSource = null;
-                            LoadTitleAllSongs();
+                            await LoadTitleAllSongs();
                             FavoriteListBox.ItemsSource = null;
                             await LoadFavoriteList();
                         }
@@ -490,11 +490,18 @@ namespace MusicPlayList
 
                         if (result == MessageBoxResult.Yes)
                         {
-                            await _favoriteService.RemoveFavoriteAsync(CurrentUser.UserId, selectedSong.SongId.Value);
+                            if (selectedSong.SongId.HasValue)
+                            {
+                                await _favoriteService.RemoveFavoriteAsync(CurrentUser.UserId, selectedSong.SongId.Value);
 
-                            // Tải lại danh sách yêu thích
-                            FavoriteListBox.ItemsSource = null;
-                            await LoadFavoriteList();
+                                // Tải lại danh sách yêu thích
+                                FavoriteListBox.ItemsSource = null;
+                                await LoadFavoriteList();
+                            }
+                            else
+                            {
+                                MessageBox.Show("Bài hát không có ID hợp lệ.", "Lỗi", MessageBoxButton.OK, MessageBoxImage.Error);
+                            }
                         }
                     }
                 }
@@ -508,8 +515,6 @@ namespace MusicPlayList
                 MessageBox.Show($"Đã xảy ra lỗi khi xóa bài hát: {ex.Message}", "Lỗi", MessageBoxButton.OK, MessageBoxImage.Error);
             }
         }
-
-
 
         private async void Addbtn_Click(object sender, RoutedEventArgs e)
         {
